backend/requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
python-multipart>=0.0.6
pillow>=9.5.0
torch>=2.0.1
torchvision>=0.15.2
numpy>=1.24.3
python-dotenv>=1.0.0
opencv-python-headless>=4.7.0.72
